+++ packages/giu/src/gral/constants.js
@@ -63,9 +63,6 @@ const UNICODE = {
 
 const MISC = {
   windowBorderBreathe: 5,
-  zHintBase: 50,
-  zModalBase: 50,
-  zModalStep: 10,
   zMainFloatDelta: 5,
   zErrorFloatDelta: 2,
   zNotif: 9000,