+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-
+* **Breaking**:
+    * Move 99.99% of styles to giu.css, so that they can be fully configured by the user.
+    * Remove `style` props.
+    * Remove `zIndex` props.
 * Bump babel.
 * Remove tests + related dependencies.